Mojo buffer api using `size_t` for size
This change follows upstream changes to make code more ergonomic around
buffer size and interacting with other types that already use `size_t`.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/f5d24010ac66465149c6275396ccc8570d21a94d

commit f5d24010ac66465149c6275396ccc8570d21a94d
Author: Lukasz Anforowicz <lukasza@chromium.org>
Date:   Tue Apr 9 17:05:16 2024 +0000

    `size_t` in `mojo::DataPipe[Consumer|Producer]Handle`: New `size_t` API.

    Motivation for using `size_t` instead of `uint32_t`:

    * It makes it easier to interoperate with C++ container APIs that
      typically use `size_t`.  In particular, it makes it easier to execute
      subsequent migration to `span` (`span::size` returns `size_t`).  See
      also WIP CL at https://crrev.com/c/5404364.
    * It supports the `-Wshorten-64-to-32` work. See also crbug.com/1201109
      and
      https://docs.google.com/document/d/1CTbQ-5cQjnjU8aCOtLiA7G6P0i5C6HpSDNlSNq6nl5E/edit?usp=sharing

    This CL introduces new `size_t`-based APIs in
    `//mojo/public/c/system/data_pipe.h`.  Notes:

    * To support incremental migration the old `uint32_t`-based APIs have
      been retained in `//mojo/public/cpp/system/data_pipe.h` for now.
      Follow-up CLs will migrate remaining users of the old APIs outside of
      `//mojo`.
    * This CL does *not* change the C-level APIs because of the requirement
      to retain a stable ABI (see also the comments at the top of
      `//mojo/public/c/system/thunks.h`).
    * This CL may have introduced new integer conversions - in those cases
      the CL attempts to have well-defined behavior (e.g. using
      `base::checked_cast`).  OTOH, this CL did *not* attempt to fix old
      integer operations (e.g. old `int` => `uint32_t` / `size_t`
      conversions, or old additions that may potentially overflow).

    Bug: 40284755

+++ components/ai_chat/renderer/ai_chat_resource_sniffer_throttle_unittest.cc
@@ -48,7 +48,7 @@ class MojoDataPipeSender {
   }
 
   void OnWritable(MojoResult) {
-    uint32_t sending_bytes = data_.size() - sent_bytes_;
+    size_t sending_bytes = data_.size() - sent_bytes_;
     MojoResult result = handle_->WriteData(
         data_.c_str() + sent_bytes_, &sending_bytes, MOJO_WRITE_DATA_FLAG_NONE);
     switch (result) {
@@ -170,7 +170,7 @@ class MockDelegate : public blink::URLLoaderThrottle::Delegate {
     source_body_handle_.reset();
   }
 
-  uint32_t ReadResponseBody(uint32_t size) {
+  size_t ReadResponseBody(size_t size) {
     std::vector<uint8_t> buffer(size);
     MojoResult result = destination_loader_client_.response_body().ReadData(
         buffer.data(), &size, MOJO_READ_DATA_FLAG_NONE);
